renamed components to views
