+++ src/views/marketing-page/images/PartNerd.png
